requirements.txt
rdflib
mysql-connector-python
flask
